rf: move `output_reports()` and supporting defs
From `cli/tools.py` to `tools.py`
